+++ src/base/b-dynamic-page/b-dynamic-page.ts
@@ -43,9 +43,12 @@ import type {
 
 	iDynamicPageEl,
 	KeepAliveStrategy,
-	UnsafeBDynamicPage
+	UnsafeBDynamicPage,
+	OnBeforeRemovePage
 
 } from 'base/b-dynamic-page/interface';
+
+import ScrollManager from 'base/b-dynamic-page/modules/scroll-manager';
 
 export * from 'super/i-data/i-data';
 export * from 'base/b-dynamic-page/interface';
@@ -242,7 +245,7 @@ export default class bDynamicPage extends iDynamicPage {
 		}
 
 		const
-			{unsafe, route} = this;
+			{unsafe, route, r} = this;
 
 		return new SyncPromise((r) => {
 			this.onPageChange = onPageChange(r, this.route);
@@ -263,6 +266,8 @@ export default class bDynamicPage extends iDynamicPage {
 					currentPageComponent = currentPageEl?.component?.unsafe;
 
 				if (currentPageEl != null) {
+					r.emit('beforeRemovePage', <OnBeforeRemovePage>{saveScroll: ScrollManager.saveScrollIntoAttribute});
+
 					if (currentPageComponent != null) {
 						const
 							currentPageStrategy = unsafe.getKeepAliveStrategy(currentPage, currentRoute);
@@ -300,6 +305,10 @@ export default class bDynamicPage extends iDynamicPage {
 
 					if (pageComponentFromCache != null) {
 						pageComponentFromCache.activate();
+
+						unsafe.async.requestAnimationFrame(() => {
+							ScrollManager.restorePageElementsScroll(pageElFromCache);
+						}, {label: $$.restorePageElementsScroll});
 
 						unsafe.$el?.append(pageElFromCache);
 						pageComponentFromCache.emit('mounted', pageElFromCache);