+++ src/pages/p-v4-dynamic-page1/p-v4-dynamic-page1.ts
@@ -11,7 +11,8 @@
  * @packageDocumentation
  */
 
-import iDynamicPage, { component } from 'super/i-dynamic-page/i-dynamic-page';
+import type { OnBeforeRemovePage } from 'base/b-dynamic-page/b-dynamic-page';
+import iDynamicPage, { component, watch } from 'super/i-dynamic-page/i-dynamic-page';
 
 export * from 'super/i-dynamic-page/i-dynamic-page';
 
@@ -19,4 +20,14 @@ export * from 'super/i-dynamic-page/i-dynamic-page';
  * Simple dynamic page for tests
  */
 @component()
-export default class pV4DynamicPage1 extends iDynamicPage {}
+export default class pV4DynamicPage1 extends iDynamicPage {
+	override readonly $refs!: iDynamicPage['$refs'] & {
+		horizontalScroll: HTMLUListElement;
+	};
+
+	@watch('rootEmitter:onBeforeRemovePage')
+	protected saveSliderScroll({saveScroll}: OnBeforeRemovePage): void {
+		this.console.log(this.$refs.horizontalScroll);
+		saveScroll(this.$refs.horizontalScroll);
+	}
+}